# swagger.yaml
openapi: 3.0.3
info:
  title: Flask OpenAI API
  description: API to generate text and images using OpenAI models, including RAG and chat memory functionality.
  version: 1.0.0
servers:
  - url: http://localhost:5000

paths:
  /:
    get:
      summary: Home
      description: Returns a simple Hello World message.
      tags:
        - General
      parameters:
        - in: header
          name: x-api-key
          schema:
            type: string
          required: true
          description: API key for authentication
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
                    example: "Hello, World!"

  /chat:
    post:
      summary: Chat
      description: Sends a user message to OpenAI and returns the response.
      tags:
        - Chat
      parameters:
        - in: header
          name: x-api-key
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  example: "Hello!"
      responses:
        "200":
          description: Chat response
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
                  version:
                    type: string
                    example: "0.1.0"
        "400":
          description: Bad request — missing or invalid content
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Missing or empty 'content' field"
        "401":
          description: Unauthorized — missing or invalid API key
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized: Missing or invalid API key"
  /generateImage:
    post:
      summary: Generate Image
      description: Creates an image based on the provided text input.
      tags:
        - Image
      parameters:
        - in: header
          name: x-api-key
          schema:
            type: string
          required: true
        - in: header
          name: response-type
          schema:
            type: string
            enum: [base64, image]
          required: true
          description: Format of returned image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  example: "A red elephant"
      responses:
        "200":
          description: Generated image
          content:
            application/json:
              schema:
                type: object
                properties:
                  base64:
                    type: string
                    example: "iVBORw0KGgoAAAANSUhEUgAA..."
                  version:
                    type: string
                    example: "0.1.0"

  /upload_docs:
    post:
      summary: Upload Documents
      description: Uploads a list of texts to store in FAISS for RAG retrieval.
      tags:
        - RAG
      parameters:
        - in: header
          name: x-api-key
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                texts:
                  type: array
                  items:
                    type: string
                  example: ["Lebanon is a country.", "Beirut is its capital."]
      responses:
        "200":
          description: Documents stored successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Stored 2 documents."

  /ask_rag:
    post:
      summary: Ask RAG
      description: Query the RAG system using uploaded documents.
      tags:
        - RAG
      parameters:
        - in: header
          name: x-api-key
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  example: "What is the capital of Lebanon?"
      responses:
        "200":
          description: Answer from RAG
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
                    example: "Beirut"

  /chat_rag_memory:
    post:
      summary: Chat with RAG + Memory
      description: Query RAG using context + memory for multi-turn conversations.
      tags:
        - Chat
        - RAG
      parameters:
        - in: header
          name: x-api-key
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                  example: "user123"
                query:
                  type: string
                  example: "What is the capital of Lebanon?"
      responses:
        "200":
          description: Answer from RAG with memory
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
                    example: "Beirut"